+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - Set unit for `pg_stat_monitor.pgsm_query_max_len` to bytes
 - Do not leave the `pgsm_create_view()` function after running `CREATE EXTENSION`
 - Specifying `USE_PGXS` is no longer necessary when building with make
+- Deprecate the `pgsm_track_application_names` parameter, application name now tracked always ([PG-2602](https://perconadev.atlassian.net/browse/PG-2602))
 
 ### Removed